Agregados Efectos de Sonido

+++ LikeWildGunman/src/Main.cpp
@@ -1,3 +1,9 @@
+// ----------------------------------------------------------------------------
+//	Materia: Modelos y Algoritmos I - TP 01
+//	Autor: Germán Daniel Caíno
+//	Versión: 1.0.0
+//	Fecha: 12/12/2017
+// ----------------------------------------------------------------------------
 #include <iostream>
 #include <cstdlib>
 #include <ctime>
@@ -6,7 +12,7 @@
 // ----------------------------------------------------------------------------
 int main()
 {
-	srand(time(NULL));
+	srand(static_cast<unsigned int>(time(NULL)));
 	try
 	{
 		lwgm::Game game;
